separation of data from testing script generator #326
